+++ Upgrade/extention download DCOM/pacs_dicom_extension_v7/lib/save_policy.js
@@ -1,0 +1,6 @@
+export function resolveBulkDicomSaveMode(hasWritableFolder, _requestedMode = '') {
+  if (!hasWritableFolder) {
+    throw new Error('Select a writable folder before downloading DICOM.');
+  }
+  return 'filesystem';
+}